rename types to reusable types. add an ActionType definition. add typescript to getQuestionActions
