--- src/types.tsx
@@ -1,3 +1,0 @@
-// type for an object whose properties are either unknown  
-// or too lengthy and complex to list
-export type ObjectType = { [key: string]: any }
